+++ app/src/main/AndroidManifest.xml
@@ -17,14 +17,14 @@
         tools:targetApi="31"
         android:name=".MovieTheaterApplication">
         <activity
-            android:name=".view.reservation.seat.SeatSelectionActivity"
+            android:name=".view.movies.reservation.seat.SeatSelectionActivity"
             android:exported="false" />
         <activity
-            android:name=".view.reservation.result.ReservationResultActivity"
+            android:name=".view.movies.reservation.result.ReservationResultActivity"
             android:exported="false"
             android:parentActivityName=".view.MainActivity"/>
         <activity
-            android:name=".view.reservation.ReservationActivity"
+            android:name=".view.movies.reservation.ReservationActivity"
             android:exported="false" />
         <activity
             android:name=".view.MainActivity"
