+++ app/src/main/java/woowacourse/movie/Provider.kt
@@ -1,10 +1,9 @@
 package woowacourse.movie
 
-import android.view.View
 import woowacourse.movie.data.MovieTheaterDatabase
 import woowacourse.movie.repository.TicketRepository
-import woowacourse.movie.view.reservation.seat.SeatSelectionContract
-import woowacourse.movie.view.reservation.seat.SeatSelectionPresenter
+import woowacourse.movie.view.movies.reservation.seat.SeatSelectionContract
+import woowacourse.movie.view.movies.reservation.seat.SeatSelectionPresenter
 import woowacourse.movie.view.reservelist.ReservationListContract
 import woowacourse.movie.view.reservelist.ReservationListPresenter
