+++ app/src/main/java/woowacourse/movie/view/movies/reservation/ReservationContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.reservation
+package woowacourse.movie.view.movies.reservation
 
 import woowacourse.movie.domain.model.ReservationInfo
 import woowacourse.movie.domain.model.Screening
