+++ app/src/main/java/woowacourse/movie/view/movies/reservation/ReservationPresenter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view.reservation
+package woowacourse.movie.view.movies.reservation
 
 import woowacourse.movie.domain.model.ReservationCount
 import woowacourse.movie.domain.model.ReservationInfo
