+++ app/src/main/res/layout/activity_reservation.xml
@@ -15,7 +15,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".view.reservation.ReservationActivity">
+    tools:context=".view.movies.reservation.ReservationActivity">
 
     <ImageView
         android:id="@+id/iv_reservation_poster"
