+++ app/src/main/res/layout/activity_reservation_result.xml
@@ -11,7 +11,7 @@
     android:id="@+id/activity_reservation_result"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.reservation.result.ReservationResultActivity">
+    tools:context=".view.movies.reservation.result.ReservationResultActivity">
 
     <TextView
         android:id="@+id/tv_cancel_description"
